Exceptions

+++ src/main/java/com/springboot/training/services/UserService.java
@@ -9,6 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.springboot.training.entity.User;
+import com.springboot.training.exception.UserExistsException;
+import com.springboot.training.exception.UserNotFoundException;
 import com.springboot.training.repositories.UserRepository;
 
 @Service
@@ -21,19 +23,29 @@ public class UserService {
 		return userRepo.findAll();
 	}
 	
-	public User createUser(User user) {
+	public User createUser(User user) throws UserExistsException {
+		User existingUser = userRepo.getByUsername(user.getUsername());
+		if(existingUser != null) throw new UserExistsException("User already exists");
 		return userRepo.save(user);
 	}
 	
-	public Optional<User> getUserById(Long id) {
-		return userRepo.findById(id);
+	public Optional<User> getUserById(Long id) throws UserNotFoundException {
+		Optional<User> user = userRepo.findById(id);
+		if(!user.isPresent()) { throw new UserNotFoundException("User doesn't exists");}
+		return user;
 	}
 	
-	public User updateUserById(User user , Long id) {
+	public User updateUserById(User user , Long id) throws UserNotFoundException {
+		Optional<User> optUser = userRepo.findById(id);
+		if(!optUser.isPresent()) throw new UserNotFoundException("User doesn't exist");
+		
 		user.setId(id);
 		return userRepo.save(user);
 	}
-	public void deleteUserById(Long id) {
+	public void deleteUserById(Long id) throws UserNotFoundException {
+		Optional<User> optUser = userRepo.findById(id);
+		if(!optUser.isPresent()) throw new UserNotFoundException("User doesn't exist");
+		
 		userRepo.deleteById(id);
 		 
 	}
